% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpmodel.R
\name{as_mrgsolve.qpmodel}
\alias{as_mrgsolve.qpmodel}
\title{Coerce qpmodel to mrgsolve}
\usage{
\method{as_mrgsolve}{qpmodel}(x, ...)
}
\arguments{
\item{x}{class 'qpmodel'}

\item{...}{ignored}
}
\value{
class 'mrgsolve' (character)
}
\description{
Coerces qpmodel to mrgsolve format.
}
\concept{mrgsolve}
