% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpmodel.R
\name{as_mrgsolve}
\alias{as_mrgsolve}
\title{Coerce to mrgsolve}
\usage{
as_mrgsolve(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed arguments}
}
\value{
class 'mrgsolve'
}
\description{
Coerces to mrgsolve format.  Generic, with method \code{\link{as_mrgsolve.qpmodel}}.
}
