% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpmodel.R
\name{as_nlmixr.qpmodel}
\alias{as_nlmixr.qpmodel}
\title{Coerce qpmodel to nlmixr Format}
\usage{
\method{as_nlmixr}{qpmodel}(x, ...)
}
\arguments{
\item{x}{class 'qpmodel'}

\item{...}{ignored}
}
\description{
Coerces qpmodel to nlmixr format.
}
\examples{
writeLines(as_nlmixr(qpmodel(ode = eqns(Aabs = -ka*Aabs))))
}
