% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpmodel.R
\name{as_nlmixr}
\alias{as_nlmixr}
\title{Coerce to nlmixr}
\usage{
as_nlmixr(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed arguments}
}
\value{
class 'nlmixr'
}
\description{
Coerces to nlmixr format.  Generic, with method \code{\link{as_nlmixr.qpmodel}}.
}
\examples{
example(as_nlmixr.qpmodel)
}
