% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpmodel.R
\name{as_nonmem.qpmodel}
\alias{as_nonmem.qpmodel}
\title{Coerce to NONMEM}
\usage{
\method{as_nonmem}{qpmodel}(x, ...)
}
\arguments{
\item{x}{class 'qpmodel'}

\item{...}{ignored}
}
\description{
Coerces qpmodel to NONMEM format.
}
\examples{
writeLines(as_nonmem(qpmodel(ode = eqns(Aabs = -ka*Aabs))))
}
