% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpmodel.R
\name{as_nonmem}
\alias{as_nonmem}
\title{Coerce to NONMEM}
\usage{
as_nonmem(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed arguments}
}
\value{
class 'nonmem': a file path for the NONMEM model code
}
\description{
Coerces to NONMEM format.  Generic, with method \code{\link{as_nonmem.qpmodel}}.
}
\examples{
# see methods
}
