% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpmodel.R
\name{eqn.make}
\alias{eqn.make}
\title{Make an Equation}
\usage{
eqn.make(arg)
}
\arguments{
\item{arg}{instance of class 'call'}
}
\description{
Makes a single equation of class 'eqn'.
}
\seealso{
Other eqn: 
\code{\link{eqn_}()}
}
\concept{eqn}
