% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpmodel.R
\name{eqn_}
\alias{eqn_}
\title{Convert Equation to a Call}
\usage{
eqn_(arg)
}
\arguments{
\item{arg}{an equation}
}
\description{
Converts an equation to a call.
}
\seealso{
Other eqn: 
\code{\link{eqn.make}()}
}
\concept{eqn}
