% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpmodel.R
\name{eqns.make}
\alias{eqns.make}
\title{Make Equations}
\usage{
eqns.make(...)
}
\arguments{
\item{...}{conventional arguments}
}
\description{
Makes equations. Similar to \code{\link{eqns}} but with standard evaluation of arguments.
}
\examples{
eqns.make()
}
\seealso{
Other eqns: 
\code{\link{eqns_}()},
\code{\link{eqns}()}
}
\concept{eqns}
