% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpmodel.R
\name{eqns}
\alias{eqns}
\title{Generate Equations}
\usage{
eqns(...)
}
\arguments{
\item{...}{unquoted arguments}
}
\description{
Generates equations. See vignettes.
}
\examples{
ode <- eqns(
  Aabs = -ka*Aabs,
  Acentral = ka*Aabs - (cl/V)*Acentral
)
}
\concept{eqns}
