% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpmodel.R
\name{eqns_}
\alias{eqns_}
\title{Convert Equations to Calls}
\usage{
eqns_(...)
}
\arguments{
\item{...}{equations}
}
\description{
Converts equations to calls.
}
\seealso{
Other eqns: 
\code{\link{eqns}()}
}
\concept{eqns}
