% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpmodel.R
\name{modelCreate_NM}
\alias{modelCreate_NM}
\title{Coerce to NONMEM}
\usage{
modelCreate_NM(qpmodel = qpModel(), filename = "run1.mod")
}
\arguments{
\item{qpmodel}{class 'qpmodel'}

\item{filename}{output filename}
}
\description{
Coerces model to NONMEM format
}
