% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpmodel.R
\name{modelCreate_R}
\alias{modelCreate_R}
\title{Coerce to mrgsolve}
\usage{
modelCreate_R(qpmodel = qpModel(), filename = "model.txt")
}
\arguments{
\item{qpmodel}{class 'qpmodel'}

\item{filename}{output filename}
}
\description{
Coerces model to mrgsolve format
}
