% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpmodel.R
\name{modelView}
\alias{modelView}
\title{View 'qpmodel'}
\usage{
modelView(qpmodel = qpModel())
}
\arguments{
\item{qpmodel}{an object of class 'qpmodel'}
}
\description{
Views 'qpmodel'.
}
\seealso{
Other qpmodel: 
\code{\link{qpModel}()}
}
\concept{qpmodel}
