% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpmodel.R
\name{ode}
\alias{ode}
\title{Update ODE}
\usage{
ode(ode = eqns())
}
\arguments{
\item{ode}{ondinary differential equations}
}
\description{
Updates ODEs.
}
\seealso{
Other updates: 
\code{\link{omega}()},
\code{\link{parameter}()},
\code{\link{sigma}()},
\code{\link{theta}()}
}
\concept{updates}
