% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpmodel.R
\name{omega}
\alias{omega}
\title{Update Omegas}
\usage{
omega(...)
}
\arguments{
\item{...}{second level random effects}
}
\description{
Updates omega.
}
\examples{
omega()
}
\seealso{
Other updates: 
\code{\link{ode}()},
\code{\link{parameter}()},
\code{\link{sigma}()},
\code{\link{theta}()}
}
\concept{updates}
