% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpmodel.R
\name{+.eqns}
\alias{+.eqns}
\title{Add Equations}
\usage{
\method{+}{eqns}(e1, e2)
}
\arguments{
\item{e1}{equation 1}

\item{e2}{equation 2}
}
\description{
Combines sets of equations. Where LHS matches,
the equation in the first argument is replaced with
the corresponding equation in the second argument.
}
\examples{
eqns(ka = exp(tvka+etaka)) + eqns(cl = exp(tvcl+etacl))
}
\seealso{
Other operators: 
\code{\link{+.qpmodel}()}
}
\concept{operators}
