% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpmodel.R
\name{^.eqn}
\alias{^.eqn}
\title{Exponentiate Equation}
\usage{
\method{^}{eqn}(eq1, eq2)
}
\arguments{
\item{eq1}{equation 1}

\item{eq2}{equation 2}
}
\description{
Exponentiates equation.
}
\seealso{
Other operators: 
\code{\link{*.eqns}()},
\code{\link{*.eqn}()},
\code{\link{*.qpmodel}()},
\code{\link{+.eqns}()},
\code{\link{+.eqn}()},
\code{\link{+.qpmodel}()},
\code{\link{-.eqn}()},
\code{\link{/.eqn}()}
}
\concept{operators}
