% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpmodel.R
\name{print.qpmodel}
\alias{print.qpmodel}
\title{Print 'qpmodel'}
\usage{
\method{print}{qpmodel}(x, ...)
}
\arguments{
\item{x}{qpmodel}

\item{...}{passed arguments}
}
\value{
used for side effects
}
\description{
Prints 'qpmodel'.  Calls format method.
}
\examples{
qpmodel()
}
\seealso{
Other qpmodel: 
\code{\link{qpmodel}()}
}
\concept{qpmodel}
