% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpmodel.R
\name{qp.abs.first.order}
\alias{qp.abs.first.order}
\title{First Order Absorption}
\usage{
qp.abs.first.order(theta = eqns(tvka = 1), omega = eqns(etaka = 0))
}
\arguments{
\item{theta}{fixed effects}

\item{omega}{second level random effects}
}
\description{
First order absorption.
}
\seealso{
Other models: 
\code{\link{qp.abs.MM}()},
\code{\link{qp.abs.TCAM}()},
\code{\link{qp.abs.zero.and.first.order}()},
\code{\link{qp.abs.zero.order}()},
\code{\link{qp.elim.MM}()},
\code{\link{qp.elim.clearance}()},
\code{\link{qp.elim}()},
\code{\link{qp.onecpt.iv.cl}()},
\code{\link{qp.onecpt.iv.rate}()},
\code{\link{qp.onecpt.oral.cl}()},
\code{\link{qp.onecpt.oral.rate}()},
\code{\link{qp.threecpt.iv.cl}()},
\code{\link{qp.threecpt.iv.rate}()},
\code{\link{qp.threecpt.oral.cl}()},
\code{\link{qp.threecpt.oral.rate}()},
\code{\link{qp.twocpt.iv.cl}()},
\code{\link{qp.twocpt.iv.rate}()},
\code{\link{qp.twocpt.oral.cl}()},
\code{\link{qp.twocpt.oral.rate}()}
}
\concept{models}
