% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpmodel.R
\name{qp.abs.TCAM}
\alias{qp.abs.TCAM}
\title{Transit Compartment Absorption Model}
\usage{
qp.abs.TCAM(
  algebraics = eqns(abs0 = inpt),
  theta = eqns(tvmtt = 1, tvntr = 1, tvf1 = 0.1),
  omega = eqns(etamtt = 0, etantr = 0, etaf1 = 0)
)
}
\arguments{
\item{algebraics}{algebraic equations}

\item{theta}{fixed effects}

\item{omega}{second level random effects}
}
\description{
Transit compartment absorption model (TCAM).
}
\seealso{
Other models: 
\code{\link{qp.abs.MM}()},
\code{\link{qp.abs.first.order}()},
\code{\link{qp.abs.zero.and.first.order}()},
\code{\link{qp.abs.zero.order}()},
\code{\link{qp.elim.MM}()},
\code{\link{qp.elim.clearance}()},
\code{\link{qp.elim}()},
\code{\link{qp.onecpt.iv.cl}()},
\code{\link{qp.onecpt.iv.rate}()},
\code{\link{qp.onecpt.oral.cl}()},
\code{\link{qp.onecpt.oral.rate}()},
\code{\link{qp.threecpt.iv.cl}()},
\code{\link{qp.threecpt.iv.rate}()},
\code{\link{qp.threecpt.oral.cl}()},
\code{\link{qp.threecpt.oral.rate}()},
\code{\link{qp.twocpt.iv.cl}()},
\code{\link{qp.twocpt.iv.rate}()},
\code{\link{qp.twocpt.oral.cl}()},
\code{\link{qp.twocpt.oral.rate}()}
}
\concept{models}
