% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpmodel.R
\name{qp.elim.MM}
\alias{qp.elim.MM}
\title{Michaelis-Menten Clearance}
\usage{
qp.elim.MM(
  theta = eqns(tvkmax = 1, tvkm50 = 1),
  omega = eqns(etakmax = 0, etakm50 = 0)
)
}
\arguments{
\item{theta}{fixed effects}

\item{omega}{second level random effects}
}
\description{
Micahaelis-Menten clearance model.
}
\seealso{
Other models: 
\code{\link{qp.abs.MM}()},
\code{\link{qp.abs.TCAM}()},
\code{\link{qp.abs.first.order}()},
\code{\link{qp.abs.zero.and.first.order}()},
\code{\link{qp.abs.zero.order}()},
\code{\link{qp.elim.clearance}()},
\code{\link{qp.elim}()},
\code{\link{qp.onecpt.iv.cl}()},
\code{\link{qp.onecpt.iv.rate}()},
\code{\link{qp.onecpt.oral.cl}()},
\code{\link{qp.onecpt.oral.rate}()},
\code{\link{qp.threecpt.iv.cl}()},
\code{\link{qp.threecpt.iv.rate}()},
\code{\link{qp.threecpt.oral.cl}()},
\code{\link{qp.threecpt.oral.rate}()},
\code{\link{qp.twocpt.iv.cl}()},
\code{\link{qp.twocpt.iv.rate}()},
\code{\link{qp.twocpt.oral.cl}()},
\code{\link{qp.twocpt.oral.rate}()}
}
\concept{models}
