% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpmodel.R
\name{qp.threecpt.iv.rate}
\alias{qp.threecpt.iv.rate}
\title{Three Compartment IV Rate Model}
\usage{
qp.threecpt.iv.rate(
  theta = eqns(tvke = 1, tvV = 1, tvk12 = 1, tvk21 = 1, tvk13 = 1, tvk31 = 1),
  omega = eqns(etake = 0, etaV = 0, etak12 = 0, etak21 = 0, etak13 = 0, etak31 = 0),
  sigma = eqns(add = 0, prop = 0)
)
}
\arguments{
\item{theta}{fixed effects}

\item{omega}{second level random effects}

\item{sigma}{first level random effects}
}
\description{
Three compartment IV rate model.
}
\seealso{
Other models: 
\code{\link{qp.abs.MM}()},
\code{\link{qp.abs.TCAM}()},
\code{\link{qp.abs.first.order}()},
\code{\link{qp.abs.zero.and.first.order}()},
\code{\link{qp.abs.zero.order}()},
\code{\link{qp.elim.MM}()},
\code{\link{qp.elim.clearance}()},
\code{\link{qp.elim}()},
\code{\link{qp.onecpt.iv.cl}()},
\code{\link{qp.onecpt.iv.rate}()},
\code{\link{qp.onecpt.oral.cl}()},
\code{\link{qp.onecpt.oral.rate}()},
\code{\link{qp.threecpt.iv.cl}()},
\code{\link{qp.threecpt.oral.cl}()},
\code{\link{qp.threecpt.oral.rate}()},
\code{\link{qp.twocpt.iv.cl}()},
\code{\link{qp.twocpt.iv.rate}()},
\code{\link{qp.twocpt.oral.cl}()},
\code{\link{qp.twocpt.oral.rate}()}
}
\concept{models}
