% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpmodel.R
\name{qp.threecpt.oral.cl}
\alias{qp.threecpt.oral.cl}
\title{Three Compartment Oral Clearance Model}
\usage{
qp.threecpt.oral.cl(
  theta = eqns(tvCL = 1, tvV = 1, tvQ = 1, tvV2 = 1, tvQ2 = 1, tvV3 = 1, tvka = 1),
  omega = eqns(etaCL = 0, etaV = 0, etaQ = 0, etaV2 = 0, etaQ2 = 0, etaV3 = 0, etaka =
    0),
  sigma = eqns(add = 0, prop = 0)
)
}
\arguments{
\item{theta}{fixed effects}

\item{omega}{second level random effects}

\item{sigma}{first level random effects}
}
\description{
Three compartment oral clearance model.
}
\examples{
qp.threecpt.oral.cl()
}
\seealso{
Other models: 
\code{\link{qp.abs.MM}()},
\code{\link{qp.abs.TCAM}()},
\code{\link{qp.abs.first.order}()},
\code{\link{qp.abs.zero.and.first.order}()},
\code{\link{qp.abs.zero.order}()},
\code{\link{qp.elim.MM}()},
\code{\link{qp.elim.clearance}()},
\code{\link{qp.elim}()},
\code{\link{qp.onecpt.iv.cl}()},
\code{\link{qp.onecpt.iv.rate}()},
\code{\link{qp.onecpt.oral.cl}()},
\code{\link{qp.onecpt.oral.rate}()},
\code{\link{qp.threecpt.iv.cl}()},
\code{\link{qp.threecpt.iv.rate}()},
\code{\link{qp.threecpt.oral.rate}()},
\code{\link{qp.twocpt.iv.cl}()},
\code{\link{qp.twocpt.iv.rate}()},
\code{\link{qp.twocpt.oral.cl}()},
\code{\link{qp.twocpt.oral.rate}()}
}
\concept{models}
