% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpmodel.R
\name{qpModel}
\alias{qpModel}
\title{Create a qpModel
Creates an object of class 'qpmodel', convertible
to NONMEM code or mrgsolve code.}
\usage{
qpModel(
  ode = eqns(),
  algebraic = eqns(),
  param = eqns(),
  omega = eqns(),
  sigma = eqns(),
  theta = eqns(),
  observe = eqns(),
  output = eqns(),
  global = eqns(),
  custom = eqns()
)
}
\arguments{
\item{ode}{ordinary differential equations}

\item{algebraic}{algebraic equations}

\item{param}{parameters}

\item{omega}{second level random effects}

\item{sigma}{first level random effects}

\item{theta}{fixed effects}

\item{observe}{items to include in output}

\item{output}{other output}

\item{global}{global equations}

\item{custom}{custom equations}
}
\description{
Create a qpModel
Creates an object of class 'qpmodel', convertible
to NONMEM code or mrgsolve code.
}
\examples{
qpModel()

}
\seealso{
Other qpmodel: 
\code{\link{modelView}()}
}
\concept{qpmodel}
