% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpmodel.R
\name{sigma}
\alias{sigma}
\title{Update Sigmas}
\usage{
sigma(sigma = eqns())
}
\arguments{
\item{sigma}{first level random effects}
}
\description{
Updates sigma.
}
\seealso{
Other updates: 
\code{\link{ode}()},
\code{\link{omega}()},
\code{\link{parameter}()},
\code{\link{theta}()}
}
\concept{updates}
