% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpmodel.R
\name{/.eqn}
\alias{/.eqn}
\title{Divide Equation}
\usage{
\method{/}{eqn}(e1, e2)
}
\arguments{
\item{e1}{equation 1}

\item{e2}{equation 2}
}
\description{
Divides equation.
}
\seealso{
Other operators: 
\code{\link{*.eqns}()},
\code{\link{*.eqn}()},
\code{\link{*.qpmodel}()},
\code{\link{+.eqns}()},
\code{\link{+.eqn}()},
\code{\link{+.qpmodel}()},
\code{\link{-.eqn}()},
\code{\link{^.eqn}()}
}
\concept{operators}
