% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpmodel.R
\name{theta}
\alias{theta}
\title{Update Thetas}
\usage{
theta(theta = eqns())
}
\arguments{
\item{theta}{fixed effects}
}
\description{
Updates theta.
}
\seealso{
Other updates: 
\code{\link{ode}()},
\code{\link{omega}()},
\code{\link{parameter}()},
\code{\link{sigma}()}
}
\concept{updates}
