% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpmodel.R
\name{*.eqns}
\alias{*.eqns}
\title{Multiply Equations}
\usage{
\method{*}{eqns}(eqns1, eqns2)
}
\arguments{
\item{eqns1}{equation 1}

\item{eqns2}{equation 2}
}
\description{
Multiplies equations.
}
\seealso{
Other operators: 
\code{\link{*.qpmodel}()},
\code{\link{+.eqns}()},
\code{\link{+.qpmodel}()}
}
\concept{operators}
