% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpmodel.R
\name{parameter}
\alias{parameter}
\title{Update Parameter Definitions}
\usage{
parameter(...)
}
\arguments{
\item{...}{parameters}
}
\description{
Updates parameter definitions.
}
\seealso{
Other updates: 
\code{\link{ode}()},
\code{\link{omega}()},
\code{\link{sigma}()},
\code{\link{theta}()}
}
\concept{updates}
