% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpmodel.R
\name{+.eqns}
\alias{+.eqns}
\title{Add to Equations}
\usage{
\method{+}{eqns}(eqns1, eqns2)
}
\arguments{
\item{eq1}{equation 1}

\item{eq2}{equation 2}
}
\description{
Adds to equations.
}
\seealso{
Other operators: 
\code{\link{*.eqns}()},
\code{\link{*.eqn}()},
\code{\link{*.qpmodel}()},
\code{\link{+.eqn}()},
\code{\link{+.qpmodel}()},
\code{\link{-.eqn}()},
\code{\link{/.eqn}()},
\code{\link{^.eqn}()}
}
\concept{operators}
