% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpmodel.R
\name{+.qpmodel}
\alias{+.qpmodel}
\title{Add 'qpmodel'}
\usage{
\method{+}{qpmodel}(qp1, qp2)
}
\arguments{
\item{qp1}{qpmodel 1}

\item{qp2}{qpmodel 2}
}
\description{
Adds 'qpmodel'.
}
\seealso{
Other operators: 
\code{\link{*.eqns}()},
\code{\link{*.qpmodel}()},
\code{\link{+.eqns}()}
}
\concept{operators}
