% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpmodel.R
\name{qpmodel}
\alias{qpmodel}
\title{Create a qpmodel
Creates an object of class 'qpmodel', convertible
to NONMEM code or mrgsolve code.}
\usage{
qpmodel(
  ode = eqns(),
  algebraic = eqns(),
  param = eqns(),
  omega = eqns(),
  sigma = eqns(),
  theta = eqns(),
  observe = eqns(),
  output = eqns(),
  global = eqns(),
  custom = eqns()
)
}
\arguments{
\item{ode}{ordinary differential equations}

\item{algebraic}{algebraic equations}

\item{param}{parameters}

\item{omega}{second level random effects}

\item{sigma}{first level random effects}

\item{theta}{fixed effects}

\item{observe}{items to include in output}

\item{output}{other output}

\item{global}{global equations}

\item{custom}{custom equations}
}
\description{
Create a qpmodel
Creates an object of class 'qpmodel', convertible
to NONMEM code or mrgsolve code.
}
\examples{
baseModel <- qpmodel(
  ode = eqns(
    Aabs = -ka*Aabs,
    Acentral = ka*Aabs - (cl/V)*Acentral
  ),
  param = eqns(
    ka = exp(tvka+etaka),
    cl = exp(tvcl+etacl),
    V = tvV + etaV
  ),
  theta = eqns(
    tvka = 0.2,
    tvcl = 1,
    tvV = 20
  ),
  omega = eqns(
    etaka = 0.1,
    etacl = 0.2,
    etaV = 0.3
  ),
  sigma = eqns(
    ADD = 0.2,
    PROP = 0.2
  ),
  observe = eqns(
    IPRED = Acentral/V, DV = IPRED*(1+PROP)+ADD
  ),
  output = eqns(
    IPRED, DV
  )
)

}
\seealso{
Other qpmodel: 
\code{\link{print.qpmodel}()}
}
\concept{qpmodel}
